% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_legendre.R
\name{.CalcLegendre}
\alias{.CalcLegendre}
\title{Compute Associated Legendre Functions Given Sequence of (degree, order) Indices}
\usage{
.CalcLegendre(mu)
}
\arguments{
\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\description{
Procedure that computes the associated Legendre function, \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}, given a sequence of (degree, order) indices and function argument \eqn{\mu}{mu}. This is computed via a closed-form equation.
}
\details{
The underlying equation used is: P_{n,m}(x)=(-1)^{m}\cdot 2^{n}\cdot (1-x^{2})^{m/2}\cdot \sum _{k=m}^{n}{\frac {k!}{(k-m)!}}\cdot x^{k-m}\cdot {\binom {n}{k}}{\binom {\frac {n+k-1}{2}}{n}}
}
