% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_coefficients.R
\name{.CalculateGaussCoef}
\alias{.CalculateGaussCoef}
\title{Lookup Table for Gauss coefficients g & h}
\usage{
.CalculateGaussCoef(n, m, t, wmmVersion = "derived")
}
\arguments{
\item{n}{Degree of associated Legendre function to compute}

\item{m}{Order of associated Legendre function to compute}

\item{t}{Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088}

\item{wmmVersion}{}
}
\value{
vector of Gauss coefficients, \eqn{g_{n,m}(t)} and \eqn{h_{n,m}(t)}
}
\description{
Find Gauss coefficient \eqn{g_{n,m}(t)} consistent with the World Magnetic Model.
}
\seealso{
\code{\link{.CalculateGaussCoefH}}
}
