% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geocoordinates.R
\name{.CalculateGeocentricFieldSum}
\alias{.CalculateGeocentricFieldSum}
\title{Calculate sum of geocentric field components}
\usage{
.CalculateGeocentricFieldSum(legendreTable, gaussCoef, radius, lon, latGC,
  deltaLatitude)
}
\arguments{
\item{legendreTable}{\code{data.table} modified by \code{.CalculateSchmidtLegendreDerivative}}

\item{gaussCoef}{Gauss coefficients as calculated by \code{.CalculateGaussCoef}}

\item{radius}{Radius of curvature of prime vertical at given geodetic latitude}

\item{lon}{GPS longitude}

\item{latGC}{GPS latitude, geocentric}

\item{deltaLatitude}{(Geocentric Latitude - Geodetic Latitude) in decimal degrees}
}
\description{
Calculate sum of geocentric field components
}
