% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geocoordinates.R
\name{.CalculateGeocentricFieldSummand}
\alias{.CalculateGeocentricFieldSummand}
\title{Calculate summand of geocentric field component}
\usage{
.CalculateGeocentricFieldSummand(legendreTable, radius, lon, latGC)
}
\arguments{
\item{legendreTable}{\code{data.table} modified by \code{.CalculateSchmidtLegendreDerivative}}

\item{radius}{Radius of curvature of prime vertical at given geodetic latitude}

\item{lon}{GPS longitude}

\item{latGC}{GPS latitude, geocentric}
}
\description{
Calculate summand of geocentric field component
}
