% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions_wmm.R
\name{.CalculateMagneticField}
\alias{.CalculateMagneticField}
\title{Calculate Expected Magnetic Field from WMM2015}
\usage{
.CalculateMagneticField(lon, latGD, latGC, radius, time,
  highestDegree = 12, wmmVersion = "derived")

.CalculateMagneticField(lon, latGD, latGC, radius, time,
  highestDegree = 12, wmmVersion = "derived")
}
\arguments{
\item{lon}{GPS longitude}

\item{latGD}{GPS latitude, geodetic}

\item{latGC}{GPS latitude, geocentric}

\item{radius}{Radius of curvature of prime vertical at given geodetic latitude}

\item{time}{Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088}

\item{highestDegree}{Highest degree used to compute the WMM magnetic field. Note: This is a diagnostic. \code{highestDegree} should always be 12.}

\item{timeInYears}{Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088}

\item{lon}{GPS longitude}

\item{latGD}{GPS latitude, geodetic}

\item{latGC}{GPS latitude, geocentric}

\item{radius}{Radius of curvature of prime vertical at given geodetic latitude}

\item{highestDegree}{Highest degree used to compute the WMM magnetic field. Note: This is a diagnostic. \code{highestDegree} should always be 12.}
}
\value{
Expected magnetic field from WMM2015 expressed as a vector, \eqn{m_{\lambda_t,\varphi_t,h_t,t}^{WMM}}{m_wmm(lambda_t, phi_t, h_t, t)}

Expected magnetic field from WMM2015 expressed as a vector, \eqn{m_{\lambda_t,\varphi_t,h_t,t}^{WMM}}{m_wmm(lambda_t, phi_t, h_t, t)}
}
\description{
Calculate the magnetic field for a given location and time using the fitted spherical harmonic model from the 2015 World Magnetic Model.

Calculate the magnetic field for a given location and time using the fitted spherical harmonic model from the 2015 World Magnetic Model.
}
\examples{
# This code reproduces the output of the numerical example described in WMM 2015
lon <- 240
latGD <- -80
height <- 1e5

geocentric <- .ConvertGeodeticToGeocentricGPS(latGD, height)

magField_WMM(
    lon = lon
    , latGD = latGD
    , latGC = geocentric[['latitude_GC']][1]
    , radius = geocentric[['radius_GC']][1]
    , timeInYears = 2017.5
)

## Expected output
# X = 5683.51754 95763 nT
# Y = 14808.84920 23104 nT
# Z = -50163.01336 54779 nT

## Output a/o 2017-10-03
# X = 5683.518 nT
# Y = 14808.85 nT
# Z = -50163.01 nT
# This code reproduces the output of the numerical example described in WMM 2015
lon <- 240
latGD <- -80
height <- 1e5

geocentric <- .ConvertGeodeticToGeocentricGPS(latGD, height)

.CalculateMagneticField(
    lon = lon,
    latGD = latGD,
    latGC = geocentric[['latitude_GC']][1],
    radius = geocentric[['radius_GC']][1],
    time = 2017.5,
    wmmVersion = 'WMM2015'
)

## Expected output
# X = 5683.51754 95763 nT
# Y = 14808.84920 23104 nT
# Z = -50163.01336 54779 nT

## Calculated Output
# X = 5683.518 nT
# Y = 14808.85 nT
# Z = -50163.01 nT
}
\seealso{
\code{\link{getMagneticFieldWMM}}

\code{\link{getMagneticFieldWMM}}
}
