% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_legendre.R
\name{.CalculateSchmidtLegendre}
\alias{.CalculateSchmidtLegendre}
\title{Calculate Schmidt semi-normalized Legendre function}
\usage{
.CalculateSchmidtLegendre(legendreTable)
}
\arguments{
\item{legendreTable}{\code{data.table} modified by \code{.RunLegendreProcedure}}
}
\description{
Calculate Schmidt semi-normalized Legendre function
}
