% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_legendre.R
\name{.CalculateSchmidtLegendreDerivative}
\alias{.CalculateSchmidtLegendreDerivative}
\title{Calculate mu-derivative of Schmidt semi-normalized Legendre function}
\usage{
.CalculateSchmidtLegendreDerivative(legendreTable, mu)
}
\arguments{
\item{legendreTable}{\code{data.table} modified by \code{.CalculateSchmidtLegendre}}

\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\description{
Calculate mu-derivative of Schmidt semi-normalized Legendre function
}
