% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_misc.R
\name{.CheckBlackoutZone}
\alias{.CheckBlackoutZone}
\title{Check if given horizontal intensity triggers a blackout zone}
\usage{
.CheckBlackoutZone(h)
}
\arguments{
\item{h}{horizontal intensity, \code{numeric}}
}
\value{
\code{warning} if warranted
}
\description{
Check if given horizontal intensity triggers a blackout zone
}
