% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geocoordinates.R
\name{.ConvertGeodeticToGeocentricGPS}
\alias{.ConvertGeodeticToGeocentricGPS}
\title{Convert from Geodetic to Geocentric Coordinates}
\usage{
.ConvertGeodeticToGeocentricGPS(latitudeGD, height)
}
\arguments{
\item{latitudeGD}{Geodetic latitude in decimal degrees}

\item{height}{Height in meters above ellipsoid (not mean sea level)}
}
\value{
List with first element as geocentric latitude in decimal degrees and second element as geocentric radius
}
\description{
Convert geodetic coordinates to geocentric coordinates
}
