% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_legendre.R
\name{.RunLegendreProcedure}
\alias{.RunLegendreProcedure}
\title{Compute Associated Legendre Functions Given Sequence of (degree, order) Indices}
\usage{
.RunLegendreProcedure(mu)
}
\arguments{
\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\description{
Procedure that computes the associated Legendre function, \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}, given a sequence of (degree, order) indices and function argument \eqn{\mu}{mu}. This is computed via recursive relationships for Legendre functions.
}
