% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDeclinationWMM}
\alias{getDeclinationWMM}
\title{Calculate Expected Declination from WMM2015}
\usage{
getDeclinationWMM(lon, lat, height, timeInYears)
}
\arguments{
\item{lon}{GPS longitude}

\item{lat}{GPS latitude, geodetic}

\item{height}{GPS height in meters above ellipsoid}

\item{timeInYears}{Annualized date time. E.g., 2015-02-01 = (2015 + 32/365) = 2015.088}

\item{atanSelection}{Method used to compute arctangent in the function \code{atanVersion}, expressed as a string}
}
\value{
Expected declination in radians from WMM2015
}
\description{
Convenience function that takes in geodetic GPS location and annualized time, and returns the expected declination (in radians using right-hand rule).
}
\seealso{
\code{\link{getMagneticFieldWMM}}
}
