% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P}
\alias{P}
\title{Calculate Associated Legendre Function}
\usage{
P(n, m, mu)
}
\arguments{
\item{n}{Degree of associated Legendre function to compute}

\item{m}{Order of associated Legendre function to compute}

\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\value{
\eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}, scalar
}
\description{
Recursively compute Associated Legendre Function, \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}, given degree \eqn{n}, order \eqn{m}, and argument \eqn{\mu}{mu}.
The function \code{P} calls the two recursive functions, \code{.RunLegendreProcedure} and \code{.CalculateRecursiveLegendre}.
}
