% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P_Schmidt}
\alias{P_Schmidt}
\title{Calculate Schmidt Semi-Normalized Associated Legendre Functions}
\usage{
P_Schmidt(n, m, mu)
}
\arguments{
\item{n}{Degree of associated Legendre function to compute}

\item{m}{Order of associated Legendre function to compute}

\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\value{
\eqn{\widetilde{P}_{n,m}(\mu)}{~P_{n,m}(mu)}, scalar
}
\description{
Convert associated Legendre function, \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}, to Schmidt semi-normalized associated Legendre function, \eqn{\widetilde{P}_{n,m}(\mu)}{~P_{n,m}(mu)}.
}
