% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{P_Schmidt_muDeriv}
\alias{P_Schmidt_muDeriv}
\title{Derivative of Schmidt Semi-Normalized Associated Legendre Function}
\usage{
P_Schmidt_muDeriv(n, m, mu)
}
\arguments{
\item{n}{Degree of associated Legendre function to compute}

\item{m}{Order of associated Legendre function to compute}

\item{mu}{Function argument to \eqn{P_{n,m}(\mu)}{P_{n,m}(mu)}}
}
\value{
\eqn{\frac{d\widetilde{P}_{n,m}(\mu)}{d\mu}}{d~P_{n,m}(mu)/d mu}, scalar
}
\description{
Calculate the derivative of Schmidt semi-normalized associated Legendre function, \eqn{\frac{d\widetilde{P}_{n,m}(\mu)}{d\mu}}{d~P_{n,m}(mu)/d mu}.
Note: This function is used to calculate the X' field component. It also needs to be multiplied by dMu / d(geocentered latitude). Since we'll evaluate this function with mu = sin(latitude), we need to multiply by cos(latitude)
}
