% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmm.R
\docType{package}
\name{wmm}
\alias{wmm}
\alias{wmm-package}
\title{wmm: R Implementation of World Magnetic Model}
\description{
The \code{wmm} package calculates magnetic field at a given location and time according to the World Magnetic Model.
}
\section{WMM functions}{

This package has 1 exported function: \code{\link{GetMagneticFieldWMM}}
}

